requirements.txt
Flask==1.1.2
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.16.4
oauthlib==3.1.0
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
Pillow==6.0.0
sqlparse==0.3.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
